Update PostgreSQL triggers in order to consider MultiDatastreams
Triggers updated Datastreams when Observations where added, modified or deleted, but MultiDatastreams weren't.
Added in the same triggers an IF section to do the same update on MultiDatastreams (the old code has been put into an IF section too).
note: in DELETE or UPDATE operation OBSERVED_AREA was not considered (I think for performance considerations) and are not considered in the MultiDatastream updating either.
In the update trigger there was a loop in the result of a query with a condition "equal" on the primary key of table DATASTREAMS, query that obviously can give only one result. I omitted the loop in the new DATASTREAM section, but didn't touch the Datastream one.