Improved performance of Datastream update triggers
The triggers would execute multiple separate update statements, one for
each property. This was very inefficient. Now the different updates are
gathered and send to the database in one update statement.
Also, the resultTime would be updated if null, even if the new value was
also null. Now resultTime is only updated if the value actually changes.