Improved insert performance by caching the ID of the last used FoI
If an inserted observation has a new FoI compared to previous Observations
then the observedArea of the Datastream needs to be updated. Instead of
checking the geometry of the FoI against the observedArea, we keep track
of the ID of the last used FoI, so we do not need to load additional data
to do the check.